Patch 1

+++ NoitaSeedChanger/Release.cs
@@ -1,8 +1,9 @@
-﻿using System;
+using System;
 using System.Diagnostics;
 using System.IO;
 using System.Text;
 using System.Threading;
+using System.Collections.Generic;
 
 namespace NoitaSeedChanger
 {
@@ -10,32 +11,65 @@ namespace NoitaSeedChanger
     {
         private static readonly string hashFile = "/_version_hash.txt";
 
-        private const string final = "c0ba23bc0c325a0dc06604f114ee8217112a23af";
+        private static string gameHash = "";
+        private static Dictionary<string, int[]> targets = null;
+        private const string final = "b71aff760755c178c9f09ebd1e61ddc5272ceb4b";
+        //TODO: FIXME
         private const string beta = "cac8fef90391e9409e8be422ec8322bb0b2cde2e";
-        private const string old = "3bbb44abfe5f4e08dcff1aba3160cd512f7e756c";
+        private const string old = "c0ba23bc0c325a0dc06604f114ee8217112a23af";
 
-        public static void Set()
+        //TODO: these belong in an XML file, like <Release hash="xxx"><target addr="0x12345678" /></Release> so that they are user configurable.
+
+        private static readonly int[] p_Final = new int[] {
+            0x145059C,0x14DE1E0             // final
+        };
+        //TODO: dunno the beta targets yet
+        private static readonly int[] p_Beta = new int[] {
+            0x1427B74, 0x1434BC4, 0x14C0178, 0x1432458  // beta branch
+        };
+        private static readonly int[] p_Old = new int[] {
+            0x14136D4, 0x1420798, 0x14ABCF0             // old version
+        };
+        private static readonly int[] empty = new int[] { };
+
+        static Release()
+        {
+            if (targets == null)
+                LoadTargets();
+        }
+
+        public static int[] Targets
+        {
+            get
+            {
+                int[] target = targets[gameHash];
+                if (target == null)
+                    return empty;
+                return target;
+            }
+        }
+        //TODO: load from XML
+        private static void LoadTargets()
+        {
+            targets = new Dictionary<string, int[]>();
+            targets.Add(final, p_Final);
+            targets.Add(beta, p_Beta);
+            targets.Add(old, p_Old);
+        }
+
+        public static void Set(Process game)
         {
             try
             {
-                string[] lines = File.ReadAllLines(GetHashFile(), Encoding.UTF8);
-                switch (lines[0])
+                string[] lines = File.ReadAllLines(GetHashFile(game), Encoding.UTF8);
+                gameHash = lines[0];
+
+                if (targets[gameHash].Length == 0)
                 {
-                    case final:
-                        Program.release = 0;
-                        break;
-                    case beta:
-                        Program.release = 1;
-                        break;
-                    case old:
-                        Program.release = 2;
-                        break;
-                    default:
-                        Helper.Error("Current game version not supported!");
-                        Helper.Error("Closing NSC in 10 seconds.");
-                        Thread.Sleep(10000);
-                        Process.GetCurrentProcess().Kill();
-                        break;
+                    Helper.Error("Current game version not supported!");
+                    Helper.Error("Closing NSC in 10 seconds.");
+                    Thread.Sleep(10000);
+                    Process.GetCurrentProcess().Kill();
                 }
             }
             catch (Exception e)
@@ -47,11 +81,11 @@ namespace NoitaSeedChanger
             }
 
         }
-        private static string GetHashFile()
+        private static string GetHashFile(Process game)
         {
             try
             {
-                return Path.GetDirectoryName(Program.game.MainModule.FileName) + hashFile;
+                return Path.GetDirectoryName(game.MainModule.FileName) + hashFile;
             }
             catch (Exception)
             {
